+++ CHANGES.rst
@@ -22,9 +22,12 @@ v3.0.0 (xxxx-xx-xx)
 - Add ``human_case``.
 - Add ``insert_substr``.
 - Add ``is_blank``.
+- Add ``is_bool`` as alias of ``is_boolean``.
 - Add ``is_builtin``, ``is_native``.
 - Add ``is_dict`` as alias of ``is_plain_object``.
+- Add ``is_int`` as alias of ``is_integer``.
 - Add ``is_match``.
+- Add ``is_num`` as alias of ``is_number``.
 - Add ``join`` as alias of ``implode``.
 - Add ``lines``.
 - Add ``number_format``.